requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
google-api-python-client
python-dotenv
yt-dlp